Added result harvesting functionality

+++ pymw/boinc_interface.py
@@ -1,22 +1,30 @@
 import pymw
+import threading
 import shutil
 import os
 import re
+import time
 
-class _Worker:
-    def __init__(self):
-	self._active_task = None
+class _ResultHandler(threading.Thread):
+    def __init__(self, task, sleeptime = 10):
+	threading.Thread.__init__(self)
+	self._task = task
+	self._sleeptime = sleeptime
 
-    def worker_record(self):
-	if self._active_task: task_name = str(self._active_task)
-	else: task_name = None
+    def run(self):
+	while 1:
+	    time.sleep(self._sleeptime)
+	    if os.path.exists(self._task._output_arg):
+	        self._task.task_finished()
+		break
 
 class BOINCInterface:
     def __init__(self, project_home):
 	self._project_home = project_home
 	self._project_download = project_home + "/download/"
-	self._project_upload = project_home + "/upload/"
 	self._project_templates = project_home + "/templates/"
+	self._boinc_wu_template = open("boinc_wu_template.xml").readlines()
+	self._boinc_result_template = open("boinc_result_template.xml").readlines()
 	
     def reserve_worker(self):
 	return None
@@ -24,43 +32,43 @@ class BOINCInterface:
     def execute_task(self, task, worker):
         in_file = task._input_arg.rpartition('/')[2]
 	out_file = task._output_arg.rpartition('/')[2]
-	# get input destination
+	# Get input destination
         cmd = self._project_home + "/bin/dir_hier_path " + in_file
         in_path = os.popen(cmd, "r").read().strip()
 	cmd = self._project_home + "/bin/dir_hier_path " + task._executable
 	exe_path = os.popen(cmd, "r").read().strip()
 	
-	# copy input files to download dir
+	# Copy input files to download dir
 	try:
-	    shutil.copyfile(task._executable, exe_path)
+	    if not os.path.exists(exe_path):
+    	        shutil.copyfile(task._executable, exe_path)
 	    shutil.copyfile(task._input_arg, in_path)
+	    time.sleep(0.5)
 	except IOError:
 	    print "Oops! Permission denied..."
 	    return
 	    
-	# create XML template for the wu
+	# Create XML template for the wu
 	wu_template = "pymw_wu_" + str(task._input_data) + ".xml"
 	dest = self._project_templates + wu_template
-	ln = open("boinc_wu_template.xml").readlines()
-	for i in range(len(ln)):
-	    if re.search("<PYMW_EXECUTABLE>", ln[i]):
-		ln[i] = ln[i].replace("<PYMW_EXECUTABLE>", task._executable)
-	    if re.search("<PYMW_INPUT>", ln[i]):
-		ln[i] = ln[i].replace("<PYMW_INPUT>", in_file)
-	    if re.search("<PYMW_CMDLINE>", ln[i]):
-		ln[i] = ln[i].replace("<PYMW_CMDLINE>", in_file + " " + out_file)
+	for i in range(len(self._boinc_wu_template)):
+	    if re.search("<PYMW_EXECUTABLE>", self._boinc_wu_template[i]):
+		self._boinc_wu_template[i] = self._boinc_wu_template[i].replace("<PYMW_EXECUTABLE>", task._executable)
+	    if re.search("<PYMW_INPUT>", self._boinc_wu_template[i]):
+		self._boinc_wu_template[i] = self._boinc_wu_template[i].replace("<PYMW_INPUT>", in_file)
+	    if re.search("<PYMW_CMDLINE>", self._boinc_wu_template[i]):
+		self._boinc_wu_template[i] = self._boinc_wu_template[i].replace("<PYMW_CMDLINE>", in_file + " " + out_file)
 
-	open(dest, "w").writelines(ln)
-	# create XML template for the result
+	open(dest, "w").writelines(self._boinc_wu_template)
+	# Create XML template for the result
 	result_template = "pymw_result_" + str(task._input_data) + ".xml"
 	dest = self._project_templates + result_template
-	ln = open("boinc_result_template.xml").readlines()
-	for i in range(len(ln)):
-	    if re.search("<PYMW_OUTPUT>", ln[i]):
-		ln[i] = ln[i].replace("<PYMW_OUTPUT>", out_file)
-	open(dest, "w").writelines(ln)
+	for i in range(len(self._boinc_result_template)):
+	    if re.search("<PYMW_OUTPUT>", self._boinc_result_template[i]):
+		self._boinc_result_template[i] = self._boinc_result_template[i].replace("<PYMW_OUTPUT>", out_file)
+	open(dest, "w").writelines(self._boinc_result_template)
 	
-	# call create_work
+	# Call create_work
 	cmd =  "create_work -appname pymw -wu_name pymw_" +  str(task._input_data)
 	cmd += " -wu_template templates/" +  wu_template
 	cmd += " -result_template templates/" + result_template 
@@ -69,3 +77,11 @@ class BOINCInterface:
 	os.chdir(self._project_home)
 	os.system(cmd)
 	os.chdir(cwd)
+	
+	# Wait for the results
+	tasks = []
+	result = _ResultHandler(task, 4)
+	tasks.append(result)
+	result.start()        
+	for result in tasks:
+	    result.join()
